+++ assets/js/components/ChecksResults/index.js
@@ -1,0 +1,3 @@
+import ChecksResults from './ChecksResults';
+
+export default ChecksResults;
